// packages/ui-engage/src/types.ts
export type IEngageScheduleDate = {
  type: string;
  month: string;
  day: string;
  dateTime: string;
} | null;

export interface IEngageMessenger {
  brandId: string;
  kind?: string;
  sentAs: string;
  content: string;
  rules?: IConditionsRule[];
}

export interface IEngageEmail {
  templateId?: string;
  subject: string;
  sender?: string;
  replyTo?: string;
  content: string;
  attachments?: IAttachment[];
}

export interface IEngageSms {
  from?: string;
  content: string;
  fromIntegrationId: string;
}
export interface IEngageNotification {
  title: string;
  content: string;
  isMobile: boolean;
}
// ...
export interface IEmailTemplate {
  _id: string;
  name: string;
  content: string;
}

export interface IEngageMessageDoc {
  kind?: string;
  type?: string;
  segmentIds?: string[];
  tagIds?: string[];
  customerTagIds?: string[];
  brandIds?: string[];
  customerIds?: string[];
  title: string;
  fromUserId?: string;
  method: string;
  isDraft?: boolean;
  isLive?: boolean;
  email?: IEngageEmail;
  messenger?: IEngageMessenger;
  notification?: IEngageNotification;
  cpId?: string;
  scheduleDate?: IEngageScheduleDate;
  shortMessage?: IEngageSms;
}